UPDATED messageHandler.py
Ajustes de messageHandler.py

+++ messageHandler.py
@@ -518,8 +518,10 @@ class MessageHandler:
                     async for message in channel.history(limit=None):
                         print("Mensaje escaneado "\
                             f"{message.content}")
-                        await message.delete()
-                        await asyncio.sleep(1)
+                        if any(reaction.emoji.name == ':x:' for
+                               reaction in message.reactions):
+                            await message.delete()
+                            await asyncio.sleep(1)
                 else:
                     await self.__send( "**Avalon-bot** no dispone de "\
                                        "permisos para "\
